refactor(utils): 统一图表生成函数并优化MinIO处理逻辑
将`generateChartImageForTool`合并到`generateChartImage`中，简化代码结构
将MinIO相关逻辑抽离到单独模块
更新所有工具文件使用新的统一函数
优化README文档并移除过时的TROUBLESHOOTING文件
