+++ src/tools/gauge.ts
@@ -1,6 +1,6 @@
 import type { EChartsOption, SeriesOption } from "echarts";
 import { z } from "zod";
-import { generateChartImageForTool } from "../utils";
+import { generateChartImage } from "../utils";
 import {
   HeightSchema,
   ThemeSchema,
@@ -40,7 +40,7 @@ export const generateGaugeChartTool = {
     title: TitleSchema,
     width: WidthSchema,
   }),
-  run: (params: {
+  run: async (params: {
     data: Array<{ name: string; value: number }>;
     height: number;
     max?: number;
@@ -125,13 +125,13 @@ export const generateGaugeChartTool = {
       },
     };
 
-    return generateChartImageForTool(
+    return await generateChartImage(
       echartsOption,
       width,
       height,
       theme,
       "png",
-      "generategauge_chart",
+      "generate_gauge_chart",
     );
   },
 };
