+++ src/tools/sankey.ts
@@ -1,6 +1,6 @@
 import type { EChartsOption, SeriesOption } from "echarts";
 import { z } from "zod";
-import { generateChartImageForTool } from "../utils";
+import { generateChartImage } from "../utils";
 import {
   HeightSchema,
   ThemeSchema,
@@ -40,7 +40,7 @@ export const generateSankeyChartTool = {
     title: TitleSchema,
     width: WidthSchema,
   }),
-  run: (params: {
+  run: async (params: {
     data: Array<{ source: string; target: string; value: number }>;
     height: number;
     nodeAlign?: "left" | "right" | "justify";
@@ -103,13 +103,13 @@ export const generateSankeyChartTool = {
       },
     };
 
-    return generateChartImageForTool(
+    return await generateChartImage(
       echartsOption,
       width,
       height,
       theme,
       "png",
-      "generatesankey_chart",
+      "generate_sankey_chart",
     );
   },
 };
