+++ src/utils/index.ts
@@ -2,7 +2,6 @@ export { zodToJsonSchema } from "./schema";
 export { renderECharts } from "./render";
 export {
   generateChartImage,
-  generateChartImageForTool,
   type ImageOutputFormat,
   type MCPContent,
   type ImageHandlerResult,
